reorganize curation module into separate files for clarity. Refactor naming.  Add in per-atom versions of dipole, quadrupoles and octupoles
