+++ modelforge-curate/modelforge/curate/__init__.py
@@ -1,1 +1,5 @@
 from ._version import __version__
+
+__all__ = ["__version__"]
+
+from .curate import Record, SourceDataset
